Moved print functions to be within their respective class. ex - print_region is now in region class. Added a print function for vtr::Rect
