+++ vpr/src/base/partition.h
@@ -53,4 +53,7 @@ class Partition {
     PartitionRegion part_region; ///< the union of regions that the partition can be placed in
 };
 
+///@brief used to print data from a Partition
+void print_partition(FILE* fp, Partition part);
+
 #endif /* PARTITION_H */
