+++ vpr/src/base/vpr_constraints.cpp
@@ -1,4 +1,5 @@
 #include "vpr_constraints.h"
+#include "partition.h"
 
 void VprConstraints::add_constrained_atom(const AtomBlockId blk_id, const PartitionId part_id) {
     constrained_atoms.insert({blk_id, part_id});
@@ -56,4 +57,33 @@ PartitionRegion VprConstraints::get_partition_pr(PartitionId part_id) {
     PartitionRegion pr;
     pr = partitions[part_id].get_part_region();
     return pr;
+}
+
+void print_constraints(FILE* fp, VprConstraints constraints) {
+    Partition temp_part;
+    std::vector<AtomBlockId> atoms;
+
+    int num_parts = constraints.get_num_partitions();
+
+    fprintf(fp, "\n Number of partitions is %d \n", num_parts);
+
+    for (int i = 0; i < num_parts; i++) {
+        PartitionId part_id(i);
+
+        temp_part = constraints.get_partition(part_id);
+
+        fprintf(fp, "\npartition_id: %zu\n", size_t(part_id));
+        print_partition(fp, temp_part);
+
+        atoms = constraints.get_part_atoms(part_id);
+
+        int atoms_size = atoms.size();
+
+        fprintf(fp, "\tAtom vector size is %d\n", atoms_size);
+        fprintf(fp, "\tIds of atoms in partition: \n");
+        for (unsigned int j = 0; j < atoms.size(); j++) {
+            AtomBlockId atom_id = atoms[j];
+            fprintf(fp, "\t#%zu\n", size_t(atom_id));
+        }
+    }
 }
