+++ vpr/src/base/vpr_constraints.h
@@ -99,4 +99,7 @@ class VprConstraints {
     vtr::vector<PartitionId, Partition> partitions;
 };
 
+///@brief used to print data from a VprConstraints object
+void print_constraints(FILE* fp, VprConstraints constraints);
+
 #endif /* VPR_CONSTRAINTS_H */
